requirements.txt
ultralytics
opencv-python
pyyaml
gdown
streamlit
wget